fix : distance data type to double
Former-commit-id: b6bedeadb0465ad545264fcfcd1380fb13fb50c6

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/dto/request/AttractionCoordinateRequestDto.java
@@ -11,6 +11,6 @@ import java.math.BigDecimal;
 public class AttractionCoordinateRequestDto {
     private Double latitude;
     private Double longitude;
-    private int distance;
+    private Double distance;
     private String sorted;
 }
